backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.15.80",
    "cryptography>=46.0.4",
    "fastapi>=0.115.0",
    "flagembedding>=1.2.0",
    "gdeltdoc>=1.12.0",
    "hmmlearn>=0.3.3",
    "httpx>=0.27.0",
    "jieba>=0.42.1",
    "numpy>=1.24.0,<2.2.0",
    "openai>=1.30.0",
    "pandas>=2.0.0",
    "prophet>=1.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyjwt[crypto]>=2.11.0",
    "pymongo>=4.6.0",
    "pymupdf>=1.23.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "redis>=5.0.0",
    "ruptures>=1.1.10",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "tavily-python>=0.3.0",
    "uvicorn[standard]>=0.30.0",
    "xgboost>=2.0.0",
]
